requirements.txt
click==7.1.2
dominate==2.5.1
Flask==1.1.2
Flask-Admin==1.5.6
Flask-Bootstrap4==4.0.2
Flask-Moment==0.9.0
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector==2.2.9
passlib==1.7.2
PyMySQL==0.9.3
SQLAlchemy==1.3.17
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.1
